backend/Chat_Agent/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-settings==2.3.4
httpx==0.27.0
anthropic==0.34.2
google-genai==0.8.0
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
respx==0.21.1
loguru==0.7.2
